Move `DamageResult` type to `src/@types/DamageResult.ts`
