+++ src/field/damage-number-handler.ts
@@ -1,6 +1,6 @@
 import { addTextObject } from "#app/ui/text/text-utils";
 import { TextStyle } from "#enums/text-style";
-import type { DamageResult } from "./pokemon";
+import type { DamageResult } from "#app/@types/DamageResult";
 import type { Pokemon } from "./pokemon";
 import { HitResult } from "#enums/hit-result";
 import { formatStat, fixedNumber } from "#app/utils";
